refactor: rename result backend to result
